Add files via upload
I used HTML and flask in unison to help create a little webpage to display the digital art that I have made in my free time. It is kind of a way for me to fuse my creativity in art with my computer science major. These are small steps I am taking to enhance my coding capabilities